Try to transform obj contained hash into dict
msgpack.packb doesn't support hashmap containing object or instance and
raise Exception.
We have to improve this behavior because fluent-logger is used by blacbox application(the case of "user
isn't developer") and it's sounds familiar to pass object to log.

This patch change behavior as following:

If we fail to make packet

 - Try to transform object contained hashmap to __dict__ and retry to
   make packet with new argument 'default=str' for msgpack.packb method.

 - If retrying to make packet with data transformed fail,
   this program log traceback instead of original data without abnormary
   termination