Log traceback if msgmack.packb raise Exception
msgpack.packb doesn't support hashmap containing object or instance and
raise Exception.
We have to improve this behavior because fluent-logger is used by
blacbox application(the case of "user
isn't developer") and it's sounds familiar to pass object to log.

This patch change so that if msgmack.packb raise Exception,
log traceback instead of original data without abnormary termination